requirements.txt
anthropic
streamlit
python-dotenv
Pillow
